Rename refinery:engine generator to refinery:extension.
