+++ core/lib/generators/refinery/extension/templates/.gitignore
