+++ core/lib/generators/refinery/extension/templates/Gemfile
