+++ core/lib/generators/refinery/extension/templates/Guardfile
