+++ core/lib/generators/refinery/extension/templates/Rakefile
