+++ core/lib/generators/refinery/extension/templates/app/controllers/refinery/namespace/admin/plural_name_controller.rb.erb
