+++ core/lib/generators/refinery/extension/templates/app/controllers/refinery/namespace/plural_name_controller.rb.erb
