+++ core/lib/generators/refinery/extension/templates/app/models/refinery/namespace/singular_name.rb.erb
@@ -21,7 +21,7 @@ module Refinery
 <% else -%>
 
       # def title was created automatically because you didn't specify a string field
-      # when you ran the refinery:engine generator. <3 <3 Refinery CMS.
+      # when you ran the refinery:extension generator. <3 <3 Refinery CMS.
       def title
         "Override def title in vendor/extensions/<%= namespacing.underscore %>/app/models/refinery/<%= namespacing.underscore %>/<%= singular_name %>.rb"
       end
