+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/admin/plural_name/_form.html.erb
