+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/admin/plural_name/_locale_picker.html.erb
