+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/admin/plural_name/_plural_name.html.erb
