+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/admin/plural_name/_singular_name.html.erb
