+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/admin/plural_name/_sortable_list.html.erb
