+++ core/lib/generators/refinery/extension/templates/app/views/refinery/namespace/plural_name/show.html.erb
