+++ core/lib/generators/refinery/extension/templates/config/locales/cs.yml
