+++ core/lib/generators/refinery/extension/templates/config/locales/en.yml
