+++ core/lib/generators/refinery/extension/templates/config/locales/es.yml
