+++ core/lib/generators/refinery/extension/templates/config/locales/fr.yml
