+++ core/lib/generators/refinery/extension/templates/config/locales/nb.yml
