+++ core/lib/generators/refinery/extension/templates/config/locales/sk.yml
