+++ core/lib/generators/refinery/extension/templates/config/locales/zh-CN.yml
