+++ core/lib/generators/refinery/extension/templates/config/routes.rb.erb
