+++ core/lib/generators/refinery/extension/templates/db/migrate/1_create_namespace_plural_name.rb.erb
