+++ core/lib/generators/refinery/extension/templates/db/seeds.rb.erb
