+++ core/lib/generators/refinery/extension/templates/lib/generators/refinery/extension_plural_name_generator.rb.erb
