+++ core/lib/generators/refinery/extension/templates/lib/refinery/plural_name.rb.erb
