+++ core/lib/generators/refinery/extension/templates/lib/refinery/plural_name/engine.rb.erb
