+++ core/lib/generators/refinery/extension/templates/lib/refinerycms-extension_plural_name.rb.erb
