+++ core/lib/generators/refinery/extension/templates/lib/tasks/refinery/extension_plural_name.rake
