+++ core/lib/generators/refinery/extension/templates/readme.md
