+++ core/lib/generators/refinery/extension/templates/refinerycms-extension_plural_name.gemspec
