+++ core/lib/generators/refinery/extension/templates/script/rails
