+++ core/lib/generators/refinery/extension/templates/spec/models/refinery/namespace/singular_name_spec.rb.erb
