+++ core/lib/generators/refinery/extension/templates/spec/requests/refinery/namespace/admin/plural_name_spec.rb.erb
