+++ core/lib/generators/refinery/extension/templates/spec/spec_helper.rb
