+++ core/lib/generators/refinery/extension/templates/spec/support/factories/refinery/plural_name.rb.erb
