+++ core/lib/generators/refinery/extension/templates/tasks/rspec.rake
