+++ core/lib/generators/refinery/extension/templates/tasks/testing.rake
