+++ core/spec/lib/generators/refinery/extension/extension_generator_with_i18n_spec.rb
@@ -1,15 +1,15 @@
 require 'spec_helper'
 require 'generator_spec/test_case'
-require 'generators/refinery/engine/engine_generator'
+require 'generators/refinery/extension/extension_generator'
 
 module Refinery
-  describe EngineGenerator do
+  describe ExtensionGenerator do
     include GeneratorSpec::TestCase
     destination File.expand_path("../../../../../../tmp", __FILE__)
 
     before do
       prepare_destination
-      run_generator %w{ rspec_product_test title:string description:text image:image brochure:resource }
+      run_generator %w{ rspec_product_test title:string description:text image:image brochure:resource --i18n title description }
     end
 
     specify do
@@ -79,14 +79,6 @@ module Refinery
           end
         end
       }
-    end
-
-    describe "attr_accessible" do
-      it "adds attributes to the list" do
-        File.open("#{destination_root}/vendor/extensions/rspec_product_tests/app/models/refinery/rspec_product_tests/rspec_product_test.rb") do |file|
-          file.grep(%r{attr_accessible :title, :description, :image_id, :brochure_id, :position}).count.should eq(1)
-        end
-      end
     end
   end
 end
